Blocks aren't interactable

+++ src/main/java/xyz/nasasupercomputer/worldclaimerutils/EventHandlers/ServerEventHandler.java
@@ -3,6 +3,8 @@ package xyz.nasasupercomputer.worldclaimerutils.EventHandlers;
 import java.awt.Component;
 import java.util.ArrayList;
 
+import javax.json.stream.JsonParser.Event;
+
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.multiplayer.chat.LoggedChatMessage.Player;
 import net.minecraft.network.chat.ChatType;
@@ -13,9 +15,11 @@ import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.block.Block;
 import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.event.entity.living.LivingDamageEvent;
 import net.minecraftforge.event.entity.player.EntityItemPickupEvent;
+import net.minecraftforge.event.entity.player.PlayerInteractEvent;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.eventbus.api.Event.Result;
@@ -58,6 +62,53 @@ public class ServerEventHandler {
 							// Dont pickup
 							event.setCanceled(true);
 							event.setResult(Result.DENY);
+							
+							if (ForgeConfigs.instantlyKillPlayer) {
+								player.kill(); // welp shouldn't have done that
+							}
+						}
+					}
+				}
+			}
+		}
+	}
+	
+	// Whenever the player right-clicks a block, air or item.
+	@SubscribeEvent
+	public static void onPlayerInteract(PlayerInteractEvent.RightClickBlock event) {
+		net.minecraft.world.entity.player.Player player = event.getEntity();
+		ItemStack item = event.getItemStack();
+		Block block = event.getLevel().getBlockState(event.getPos()).getBlock();
+		
+		if (ForgeConfigs.enableMod) { 
+			for (String index : ForgeConfigs.modBans) {
+				String[] fullList = index.split(",");
+				String bannedUsername = fullList[0];
+				
+				// If the players name is mentioned
+				if (player.getName().getString().equalsIgnoreCase(bannedUsername)) {
+	
+					// Loop through all the modids and see if they match the 
+					for (int i = 0; i < fullList.length - 1; i++) { // -1 to account for the first element being the players username
+						
+						// Check for the held item
+						if (ForgeRegistries.ITEMS.getKey(item.getItem()).getNamespace().equalsIgnoreCase(fullList[i + 1])) {
+							// Dont interact
+							event.setCanceled(true);
+							event.setResult(Result.DENY);
+							event.setUseBlock(Result.DENY);
+							
+							if (ForgeConfigs.instantlyKillPlayer) {
+								player.kill(); // welp shouldn't have done that
+							}
+						}
+						
+						// Check for the clicked block
+						else if (ForgeRegistries.ITEMS.getKey(block.asItem()).getNamespace().equalsIgnoreCase(fullList[i + 1])) {
+							// Dont interact
+							event.setCanceled(true);
+							event.setResult(Result.DENY);
+							event.setUseBlock(Result.DENY);
 							
 							if (ForgeConfigs.instantlyKillPlayer) {
 								player.kill(); // welp shouldn't have done that
